Handle adverse connection conditions on resume
It appears that if WinNUT resumes operation from a suspended state and the computer's networking is not prepared for a connection, then WinNUT will only try reconnecting once and hard-fail without launching a reconnection timer. Connect_UPS has been modified so that it will launch the reconnection timer if required by the caller.
- Reconnection interval is now 5 seconds regardless of compiler constants