Attempt at fixing weird log file locations
Trying to fix log files ending up in the C:\ directory. SetupAppDirectory is now private, and named GetAppDirectory. This is now strictly a function that tries to create the base directory and return it. It will also fallback to the temp directory if necessary, and hopefully log the error. The Globals method has also generally been rearranged and cleaned up with respect to logging and directories.