Fix unable to exit with tray closing functionality
When Minimize and/or Close to tray is enabled, any calls to Application.Exit are cancelled and WinNUT instead simply hides its self as if it were being minimized/closed to the tray. The application now checks if the user meant to minimize/close it to the tray, otherwise it will exit normally.