Fix ListVarGui unable to display certain nodes
The node tree is created with the root node having the name of the UPS. Since UPS variables do not have the UPS name at the beginning, the UPS name had to be stripped from the node's fullpath. The Replace function was previously unlimited, which caused it to strip out the UPS name wherever it appeared in the node path. If a UPS's name was 'ups', then any variable in a 'ups' group would become unreadable. The Replace function now only removes the UPS name from the beginning of the full path.