New power calculation methods
Replacing old power calculation methods with new system. On connection, WinNUT determines what power/load data variables are available and sets an enum variable accordingly. If no such method is available, a warning is logged to the output, and future attempts to get the power value will return with 0. On every update, the appropriate method is selected from the enum variable and the calculation is returned. The hope is that this covers the majority of UPS devices.
- Also cleaned up or rearranged code in the UPS_Device class.