Fix 'nominal in frequency' setting reading in Pref_Gui
In the old Registry preferences storage system, the 'frequency in nominal' value was stored as an index of the `Cbx_Freq_Input` combobox control in the `Pref_Gui` preferences form. However, the new Settings system stores the frequency value literally. To work around this, we set the combobox selectedindex value to the first full match with the loaded Settings value. When saving, the value of the selecteditem is read directly into the Settings value.