Upgrade Settings at start if necessary
Application Settings upgrade routine is only run automatically when using the ClickOnce installer. Since it's not run with our current MSI build, this code will then run when the program is started for the first time (nothing happens) or if it's an upgraded MSI installation with a fresh copy of settings. In that case, the Upgrade function will find the older copy and import/apply them. This operation won't be attempted until the next upgrade. Need to remove once using pure ClickOnce.